+++ cli/srecli/group/__init__.py
@@ -553,6 +553,42 @@ def build_limit_option(**kwargs) -> Callable:
     return click.option('--limit', '-l', **params)
 
 
+def build_dojo_product_option(**kwargs) -> Callable:
+    params = dict(
+        type=str, required=False,
+        help='Defect Dojo product name to which the results will be '
+             'uploaded in case Defect Dojo integration is configured. '
+             '"tenant_name", "customer_name" and "job_id" can be used '
+             'as generic placeholders inside curly brackets'
+    )
+    params.update(**kwargs)
+    return click.option('--dojo_product', '-dp', **params)
+
+
+def build_dojo_engagement_option(**kwargs) -> Callable:
+    params = dict(
+        type=str, required=False,
+        help='Defect Dojo engagement name to which the results will be '
+             'uploaded in case Defect Dojo integration is configured. '
+             '"tenant_name", "customer_name" and "job_id" can be used '
+             'as generic placeholders inside curly brackets'
+    )
+    params.update(**kwargs)
+    return click.option('--dojo_engagement', '-de', **params)
+
+
+def build_dojo_test_option(**kwargs) -> Callable:
+    params = dict(
+        type=str, required=False,
+        help='Defect Dojo test title to which the results will be '
+             'uploaded in case Defect Dojo integration is configured. '
+             '"tenant_name", "customer_name" and "job_id" can be used '
+             'as generic placeholders inside curly brackets'
+    )
+    params.update(**kwargs)
+    return click.option('--dojo_test', '-dt', **params)
+
+
 tenant_option = build_tenant_option()
 account_option = build_account_option()
 
@@ -579,3 +615,7 @@ exception_expire_at_required_option = build_iso_date_option(
 limit_option = build_limit_option()
 next_option = click.option('--next_token', '-nt', type=str, required=False,
                            help='Token to start record-pagination from')
+
+dojo_product_option = build_dojo_product_option()
+dojo_engagement_option = build_dojo_engagement_option()
+dojo_test_option = build_dojo_test_option()